refactor(chrome-extension): use directly chrom sync instances
