+++ apps/chrome-extension/src/shared/storage/index.ts
@@ -1,3 +1,2 @@
 export { Storage, StorageItems } from './storage'
-export { MemoryStorage } from './types/memoryStorage'
-export { ChromeSyncStorage } from './types/chromeSyncStorage'
+export { MemoryStorage } from './memoryStorage'
