--- apps/chrome-extension/src/shared/storage/types/chromeSyncStorage.ts
@@ -1,11 +1,0 @@
-import { Storage, StorageItems, StorageKeys } from '../storage'
-
-export class ChromeSyncStorage implements Storage {
-  get(keys?: StorageKeys): Promise<StorageItems> {
-    return chrome.storage.sync.get(keys)
-  }
-
-  async set(items: StorageItems): Promise<void> {
-    await chrome.storage.sync.set(items)
-  }
-}
